requirements.txt
# PERMA-V Wellbeing Project Requirements
# Note: This file replaces both requirements.txt and requirement.txt

# Core dependencies
fastapi>=0.110.0
pydantic==2.11.3
pydantic_core==2.33.1
uvicorn==0.34.0
mcp==1.6.0
typing_extensions==4.13.2
pytz
flask
# HTTP and API dependencies
requests==2.32.3
httpx==0.28.1
httpcore==1.0.8
anyio==4.9.0
starlette==0.46.1
sse-starlette==2.2.1
httpx-sse==0.4.0

# CLI and formatting dependencies
typer==0.15.2
rich==14.0.0
click==8.1.8
shellingham==1.5.4
Pygments==2.19.1
markdown-it-py==3.0.0
mdurl==0.1.2

# Utilities and support packages
python-dotenv==1.1.0
annotated-types==0.7.0
pydantic-settings==2.8.1
sniffio==1.3.1
h11==0.14.0
idna==3.10
exceptiongroup==1.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
urllib3==2.0.7
typing-inspection==0.4.0 